+++ app/models/service_related_platform.rb
@@ -4,6 +4,6 @@ class ServiceRelatedPlatform < ApplicationRecord
   belongs_to :service
   belongs_to :platform
 
-  validates :service, presence: true, uniqueness: { scope: :service_id }
+  validates :service, presence: true, uniqueness: { scope: :platform_id }
   validates :platform, presence: true
 end